+++ tgui/packages/tgui/interfaces/MedicalRecords/MedicalRecordsList.tsx
@@ -1,27 +1,16 @@
-import { useState } from 'react';
-
 import { useBackend } from '../../backend';
 import { Box, Button, Input } from '../../components';
-import { selectRecords } from './functions';
-import { Data, record } from './types';
+import { Data } from './types';
 
-export const GeneralRecordsList = (props) => {
+export const MedicalRecordsList = (props) => {
   const { act, data } = useBackend<Data>();
-
-  const [searchText, setSearchText] = useState<string>('');
-
-  const records: record[] = selectRecords(data.records!, searchText);
+  const { records } = data;
   return (
     <>
-      <Box mb="0.2rem">
-        <Button icon="pen" onClick={() => act('new')}>
-          New Record
-        </Button>
-      </Box>
       <Input
         fluid
         placeholder="Search by Name, DNA, or ID"
-        onInput={(e, value: string) => setSearchText(value)}
+        onChange={(e, value: string) => act('search', { t1: value })}
       />
       <Box mt="0.5rem">
         {records.map((record, i) => (
