+++ tgui/packages/tgui/interfaces/MedicalRecords/constants.ts
@@ -1,0 +1,7 @@
+export const severities = {
+  Minor: 'good',
+  Medium: 'average',
+  'Dangerous!': 'bad',
+  Harmful: 'bad',
+  'BIOHAZARD THREAT!': 'bad',
+};
